+++ .github/workflows/deploy-azure-webapps.yml
@@ -62,11 +62,11 @@ jobs:
 
       - name: Package CallWithChat Sample Artifact
         run: rushx package
-        working-directory: ./samples/Meeting
+        working-directory: ./samples/CallWithChat
 
       - name: 'Deploy CallWithChat Sample WebApp'
         uses: azure/webapps-deploy@v2
         with:
           app-name: acs-ui-dev-web-meeting-hero
           publish-profile: ${{ secrets.AZURE_MEETING_WEBAPP_PUBLISH_PROFILE }}
-          package: ./samples/Meeting/dist
+          package: ./samples/CallWithChat/dist
