+++ packages/react-composites/package.json
@@ -18,17 +18,17 @@
     "build:e2e": "rushx _by-flavor \"rushx build:e2e:call && rushx build:e2e:chat && rushx build:e2e:callwithchat\"",
     "build:e2e:call": "rushx _by-flavor \"webpack-cli build --config ./tests/browser/call/app/webpack.config.js --mode=production --env production\"",
     "build:e2e:chat": "rushx _by-flavor \"webpack-cli build --config ./tests/browser/chat/app/webpack.config.js --mode=production --env production\"",
-    "build:e2e:callwithchat": "rushx _by-flavor \"webpack-cli build --config ./tests/browser/meeting/app/webpack.config.js --mode=production --env production\"",
+    "build:e2e:callwithchat": "rushx _by-flavor \"webpack-cli build --config ./tests/browser/callwithchat/app/webpack.config.js --mode=production --env production\"",
     "test": "rushx _by-flavor \"npm run _test:by-flavor\"",
     "test:coverage": "rushx _by-flavor \"npm run test:_by-flavor -- --coverage\"",
     "test:e2e": "rushx _by-flavor \"rushx test:e2e:call && rushx test:e2e:chat && rushx test:e2e:callwithchat\"",
     "test:e2e:call": "rushx _by-flavor \"playwright test tests.browser.call\"",
     "test:e2e:chat": "rushx _by-flavor \"playwright test tests.browser.chat\"",
-    "test:e2e:callwithchat": "rushx _by-flavor \"playwright test tests.browser.meeting\"",
+    "test:e2e:callwithchat": "rushx _by-flavor \"playwright test tests.browser.callwithchat\"",
     "test:e2e:update": "rushx _by-flavor \"rushx test:e2e:call:update && rushx test:e2e:chat:update && rushx test:e2e:callwithchat:update\"",
     "test:e2e:call:update": "rushx _by-flavor \"playwright test tests.browser.call --update-snapshots\"",
     "test:e2e:chat:update": "rushx _by-flavor \"playwright test tests.browser.chat --update-snapshots\"",
-    "test:e2e:callwithchat:update": "rushx _by-flavor \"playwright test tests.browser.meeting --update-snapshots\"",
+    "test:e2e:callwithchat:update": "rushx _by-flavor \"playwright test tests.browser.callwithchat --update-snapshots\"",
     "snapshot:update": "npm run test -- --update-snapshot",
     "preprocess": "cpx \"./src/**\" ./preprocessed && babel ./src --out-dir ../preprocessed --extensions \".ts,.tsx\" --keep-file-extension --config-file ./.babelrc.js --relative && rimraf ../preprocessed",
     "api-extractor": "rushx _by-flavor \"npm run build:esm && npm run _api-extractor:by-flavor\"",
