+++ packages/react-composites/src/composites/CallWithChatComposite/CallWithChatControlBar.tsx
@@ -8,8 +8,8 @@ import { PeopleButton } from './PeopleButton';
 import { concatStyleSets, IStyle, ITheme, mergeStyles, Stack, useTheme } from '@fluentui/react';
 import { controlBarContainerStyles } from '../CallComposite/styles/CallControls.styles';
 import { callControlsContainerStyles } from '../CallComposite/styles/CallPage.styles';
-import { CallWithChatControlOptions } from './MeetingComposite';
-import { useCallWithChatCompositeStrings } from './hooks/useMeetingCompositeStrings';
+import { CallWithChatControlOptions } from './CallWithChatComposite';
+import { useCallWithChatCompositeStrings } from './hooks/useCallWithChatCompositeStrings';
 import { ChatAdapter } from '../ChatComposite';
 import { ChatButtonWithUnreadMessagesBadge } from './ChatButtonWithUnreadMessagesBadge';
 import { BaseCustomStyles, ControlBarButtonStyles } from '@internal/react-components';
