+++ samples/CallWithChat/public/index.html
@@ -6,8 +6,8 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="description" content="Azure Communication Services - UI Library: Meeting Sample app" />
-    <title>UI Library Meeting Sample</title>
+    <meta name="description" content="Azure Communication Services - UI Library: Call With Chat Sample app" />
+    <title>UI Library Call With Chat Sample</title>
   </head>
 
   <body>
