+++ samples/CallWithChat/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "name": "Azure Communication Meeting Sample",
-  "short_name": "AzureCommunicationMeetingSample",
+  "name": "Azure Communication Call with Chat Sample",
+  "short_name": "AzureCommunicationCallWithChatSample",
   "start_url": ".",
   "display": "standalone",
   "theme_color": "#000000",
