+++ samples/CallWithChat/src/app/App.tsx
@@ -20,7 +20,7 @@ import {
   getTeamsLinkFromUrl,
   isOnIphoneAndNotSafari
 } from './utils/AppUtils';
-import { CallWithChatScreen } from './views/MeetingScreen';
+import { CallWithChatScreen } from './views/CallWithChatScreen';
 import { HomeScreen } from './views/HomeScreen';
 import { UnsupportedBrowserPage } from './views/UnsupportedBrowserPage';
 import { getEndpointUrl } from './utils/getEndpointUrl';
